refactor: 공통 환경 조성을 위한 리펙토링
