+++ client/src/components/common/ChatSideBar.tsx
