+++ client/src/components/common/ProfilePhoto.tsx
