+++ client/src/components/common/SideBar.tsx
