+++ client/src/components/index.tsx
@@ -1,9 +1,9 @@
-import AlertSideBar from './AlertSideBar';
-import ChatSideBar from './ChatSideBar';
-import GroupSideBar from './GruopSideBar';
-import InfoSideBar from './InfoSideBar';
-import Gnb from './Gnb';
-import SideBar from './SideBar';
-import ProfilePhoto from './ProfilePhoto';
+import AlertSideBar from './common/AlertSideBar';
+import ChatSideBar from './common/ChatSideBar';
+import GroupSideBar from './common/GruopSideBar';
+import InfoSideBar from './common/InfoSideBar';
+import Gnb from './common/Gnb';
+import SideBar from './common/SideBar';
+import ProfilePhoto from './common/ProfilePhoto';
 
 export { AlertSideBar, ChatSideBar, GroupSideBar, InfoSideBar, Gnb, SideBar, ProfilePhoto };
