+++ client/src/pages/index.tsx
@@ -1,3 +1,4 @@
 import HomePage from './HomePage';
+import LoginPage from './LoginPage';
 
-export { HomePage };
+export { HomePage, LoginPage };
